improve layout in profile, support scroll back to previous page

+++ backend/src/controllers/authController.js
@@ -99,6 +99,9 @@ exports.deleteUser = async (req, res) => {
         const token_id = req.user.userId;
         if (token_id !== user_id) return res.status(401).json({ error: "Invalid User ID" });
 
+        const role = req.user.role;
+        if (role === "admin") return res.status(403).json({ error: "Admin account cannot delete" });
+
         await deleteUserById(user_id);
         res.status(200).json({message: "User deleted successfully"});
     } catch (err) {
